Making "portrait" type cover square screens
The idea of this npm module is to replicate mq-scss in js. mq-scss uses orientation media queries for orientation which do not feature a "square" value. According to chrome, square screens are covered under the "portrait" orientation type.

In light of that infromation, I am removing the "square" orientation type and setting the "portrait" tyoe to width <= height rather than width < height.