Adding if statement examples to the readme
I think mq-js isn't as popular as it should be due to the readme not stating that the functions can be used in if statements.

I'm adding if statement documentation to the readme now so hopefully that improves it's popularity.

mq-scss is becoming super popular with 94 stars on GitHub and a fairly large number of downloads every day. mq-js on the other hand only has 3 stars and is lucky if it gets even just 1 download in a day. It doesn't make sense when I have mentioned mq-js in the mq-scss documentation intro. Yeah js media queries aren't as essential as css media queries but I don't think that's what the problem is here.