[breaking change] screenSize().ratio now returns an object
The object holds a number and a string version of the ratio.

I thought it might be nice to have the option of which one you would like to use out of the number or the string format for the ratio.

I'm checking if a callback exists before running the callback code since screenSize() is a bit less performant now.