Accordion style content proof of concept
I think this style makes it a bit easier to digest the information and find whatr you are looking for.

It cuts the content into small, easily digestible chuncks rather than giving all of it to you all at once.

Needs styling though (I think mq-js is instore for a design overhaul 😉)